feat: notificaciones angular

+++ front/ecommerce-v1-frontend/src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/common/product';
 import { ProductService } from 'src/app/services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -28,9 +29,30 @@ export class ProductListComponent implements OnInit{
   }
 
   deleteProductById(id:string){
-    this.productService.deleteProductById(id).subscribe(
-      () => this.listProducts() 
-    );
+
+    Swal.fire({//libreria sweetalert2
+      title: "Estas seguro?",
+      text: "Una vez borrado no podras recuperarlo!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si, borrarlo!",
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteProductById(id).subscribe(
+          () => this.listProducts() 
+        );
+        Swal.fire({
+          title: "Listo, Borrado!",
+          text: "Producto eliminado",
+          icon: "success"
+        });
+      }
+    });
+
+   
   }
 
 }
